% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equi7grid-package.R
\docType{package}
\name{equi7grid-package}
\alias{equi7grid}
\alias{equi7grid-package}
\title{equi7grid: Use 'Equi7Grid'}
\description{
Provides helpers for 'Equi7Grid' a global system of 7 parts for regionalizing high resolution satellite imagery.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mdsumner/equi7grid}
  \item Report bugs at \url{https://github.com/mdsumner/equi7grid/issues}
}

}
\author{
\strong{Maintainer}: Michael Sumner \email{mdsumner@gmail.com}

}
\keyword{internal}
