% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equi7grid-package.R
\docType{package}
\name{equi7grid-package}
\alias{equi7grid}
\alias{equi7grid-package}
\title{equi7grid: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mdsumner/equi7grid}
  \item Report bugs at \url{https://github.com/mdsumner/equi7grid/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
